+++ Cargo.toml
@@ -1,2 +1,37 @@
-[workspace]
-members = ["mpesa_derive", "mpesa_core"]
+[package]
+name = "mpesa"
+version = "0.4.2"
+authors = ["Collins Muriuki <murerwacollins@gmail.com>"]
+edition = "2018"
+description = "A wrapper around the M-PESA API in Rust."
+keywords = ["api", "mpesa", "mobile"]
+repository = "https://github.com/collinsmuriuki/mpesa-rust"
+readme = "../README.md"
+license = "MIT"
+
+[dependencies]
+serde_json = "1.0"
+serde_repr = "0.1"
+dotenv = "0.15.0"
+openssl = "0.10.30"
+base64 = "0.13.0"
+failure = "0.1.5"
+failure_derive = "0.1.5"
+chrono = "0.4"
+
+[dependencies.reqwest]
+version = "0.11.3"
+features = ["blocking", "json"]
+
+[dependencies.serde]
+version = "1.0.114"
+features = ["derive"]
+
+[features]
+default = ["b2b", "b2c" ,"account_balance", "c2b_register", "c2b_simulate", "express_request"]
+b2b = []
+b2c = []
+account_balance = []
+c2b_register = []
+c2b_simulate = []
+express_request = []
