--- mpesa_derive/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "mpesa_derive"
-version = "0.3.0"
-authors = ["Collins Muriuki <murerwacollins@gmail.com>"]
-edition = "2018"
-description = "`MpesaSecurity` trait derive macro crate."
-repository = "https://github.com/collinsmuriuki/mpesa-rust"
-license = "MIT"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = "1.0"
-quote = "1.0"
