+++ src/services/c2b_register.rs
