+++ tests/c2b_register_test.rs
